reorganize the plots module
