+++ pyhgf/plots/graphviz/__init__.py
@@ -1,0 +1,11 @@
+from .plot_correlations import plot_correlations
+from .plot_network import plot_network
+from .plot_nodes import plot_nodes
+from .plot_trajectories import plot_trajectories
+
+__all__ = [
+    "plot_correlations",
+    "plot_network",
+    "plot_nodes",
+    "plot_trajectories",
+]
