+++ pyhgf/plots/networkx/__init__.py
@@ -1,0 +1,5 @@
+from .plot_network import plot_network
+
+__all__ = [
+    "plot_network",
+]
